+++ VeriTask/VeriTask/DeleteTask.cpp
@@ -2,7 +2,7 @@
 
 void DeleteTask::executeCommand(Identifier infoIdentifier, DataStorage &_dataStorage, TextUI textUI, File file) {
 	int taskNum=std::stoi(infoIdentifier.GetTaskNum());;
-	double taskIDToDelete = _dataStorage.retrieveTaskID(taskNum);
+	unsigned long long taskIDToDelete = _dataStorage.retrieveTaskID(taskNum);
 	
 	_dataStorage.deleteData(taskIDToDelete);
 	file.saveFile(_dataStorage.retrieveTaskList(), _dataStorage.getTaskIndex());
